Move KNXModule class to separate module (#146100)
