+++ homeassistant/components/knx/notify.py
@@ -1,4 +1,4 @@
-"""Support for KNX/IP notifications."""
+"""Support for KNX notify entities."""
 
 from __future__ import annotations
 
@@ -12,9 +12,9 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 from homeassistant.helpers.typing import ConfigType
 
-from . import KNXModule
 from .const import KNX_ADDRESS, KNX_MODULE_KEY
 from .entity import KnxYamlEntity
+from .knx_module import KNXModule
 
 
 async def async_setup_entry(
