+++ homeassistant/components/knx/select.py
@@ -1,4 +1,4 @@
-"""Support for KNX/IP select entities."""
+"""Support for KNX select entities."""
 
 from __future__ import annotations
 
@@ -20,7 +20,6 @@ from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType
 
-from . import KNXModule
 from .const import (
     CONF_PAYLOAD_LENGTH,
     CONF_RESPOND_TO_READ,
@@ -30,6 +29,7 @@ from .const import (
     KNX_MODULE_KEY,
 )
 from .entity import KnxYamlEntity
+from .knx_module import KNXModule
 from .schema import SelectSchema
