+++ homeassistant/components/knx/switch.py
@@ -1,4 +1,4 @@
-"""Support for KNX/IP switches."""
+"""Support for KNX switch entities."""
 
 from __future__ import annotations
 
@@ -25,7 +25,6 @@ from homeassistant.helpers.entity_platform import (
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType
 
-from . import KNXModule
 from .const import (
     CONF_INVERT,
     CONF_RESPOND_TO_READ,
@@ -35,6 +34,7 @@ from .const import (
     KNX_MODULE_KEY,
 )
 from .entity import KnxUiEntity, KnxUiEntityPlatformController, KnxYamlEntity
+from .knx_module import KNXModule
 from .schema import SwitchSchema
 from .storage.const import CONF_ENTITY, CONF_GA_SWITCH
 from .storage.util import ConfigExtractor
