+++ tests/components/knx/test_events.py
@@ -4,7 +4,8 @@ import logging
 
 import pytest
 
-from homeassistant.components.knx import CONF_EVENT, CONF_TYPE, KNX_ADDRESS
+from homeassistant.components.knx.const import KNX_ADDRESS
+from homeassistant.const import CONF_EVENT, CONF_TYPE
 from homeassistant.core import HomeAssistant
 
 from .conftest import KNXTestKit
